Initial integration with Gradle
