+++ build.gradle
@@ -1,0 +1,11 @@
+// ./build.gradle
+// This is the root level build file for Gradle
+
+repositories {
+    jcenter()
+}
+
+// Use this task to update the version of the Gradle Wrapper (gradlew)
+task wrapper(type: Wrapper) {
+    gradleVersion = '3.3'
+}
