+++ projects/team/team3729/build.properties
@@ -1,0 +1,8 @@
+# Remove this when everything is incorporated into Gradle
+
+# Project specific information
+package=org.usfirst.frc.team3729.robot
+robot.class=${package}.Robot
+simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
+#Uncomment and point at user libraries to include them in the build. Do not put libraries in the \wpilib\java folder, this folder is completely overwritten on plugin update.
+#userLibs=${user.home}/wpilib/user/lib
